feat:include data of client for generate Order,change method OrderCreate now return of the pdf

+++ back/src/main/java/com/back/chatbot/persistance/entity/ItemsOrderEntity.java
@@ -24,7 +24,7 @@ public class ItemsOrderEntity {
 
     private String name;
 
-    private BigDecimal quantity;
+    private Integer quantity;
 
     private BigDecimal price;
